requirements.txt
aiotdlib==0.22.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==1.10.15
pip==23.2.1
wheel==0.41.2
pika==1.3.2
setuptools==68.2.0
ujson==5.9.0
sortedcontainers==2.4.0
greenlet==3.0.3
typing_extensions==4.11.0
pydantic_core==2.18.1
fastapi==0.110.1
h11==0.14.0
exceptiongroup==1.2.0
sniffio==1.3.1
click==8.1.7
starlette==0.37.2
anyio==4.3.0
uvicorn==0.29.0